fix: allow null as an argument for key in config functions
The functions getValue, setValue, existValue, set, del, and save within the class Config in
src/QUI/Config.php now accept null as a valid input for the key argument. This makes the API more
flexible to use. Also, argument type for value in setValue and set and for filename in save
function has been adjusted to include null.